ai-job-finder/requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pdfminer.six==20221105
